refactor(routes): move research domain into routes/research/ subpackage
Slice 2b of the route-domain reorganization (#4082/#4071, per
specs/architecture-runtime-inventory.md §6.3). Moves research_routes.py into
routes/research/, leaving a backward-compat sys.modules shim at the old path.
Pure file reorganization, no behavior change.

The shim uses sys.modules replacement (same pattern as the merged gallery
slice #4903) so that `import routes.research_routes`, `from
routes.research_routes import X`, `importlib.import_module(...)`, and the
string-targeted `monkeypatch.setattr("routes.research_routes.DEEP_RESEARCH_DIR",
...)` in test_research_owner_scope_routes.py all operate on the same module
object the application uses.

The canonical module does NOT depend on the shim — it imports only from
core/, src/, and stdlib (research_routes.py is a leaf with zero internal
routes/ coupling).

Two source-introspection test sites repointed to the new canonical path:
- test_aux_llm_owner_scope.py
- test_endpoint_owner_scope_followup.py (shared with gallery; gallery entry
  already repointed by #4903, research entry repointed here)

Adds tests/test_research_routes_shim.py to pin the sys.modules shim contract
(legacy and canonical paths resolve to the same module object; string-targeted
monkeypatch via legacy path reaches the canonical module).

Verified: compileall clean; full suite 4210 passed, 3 skipped.
